Refactor PlatformStateRepository class
